+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Enhancements
 * The sync client will gracefully handle compensating write error messages from the server and pass detailed info to the SDK's sync error handler about which objects caused the compensating write to occur. ([#5528](https://github.com/realm/realm-core/pull/5528))
+* Support for asymmetric sync. Tables can be marked as Asymmetric when opening the realm. Upon creation, asymmetric objects are synched unidirectionally. ([#5505](https://github.com/realm/realm-core/pull/5505))
 * Close and delete all the realm files from file system in case `initialization_function` throws. Exposing this via the C API. ([#5540](https://github.com/realm/realm-core/pull/5540))
 
 ### Fixed